Added authentication on journal entry endpoints, Re-structure files

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -1,18 +1,18 @@
-package net.engineeringdigest.journalApp.journalEntry;
+package net.engineeringdigest.journalApp.service;
 
 import java.util.List;
 import java.util.Optional;
 
+import net.engineeringdigest.journalApp.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.scheduling.annotation.Async;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import net.engineeringdigest.journalApp.journalEntry.entity.JournalEntry;
-import net.engineeringdigest.journalApp.user.UserRepository;
-import net.engineeringdigest.journalApp.user.UserService;
-import net.engineeringdigest.journalApp.user.entity.User;
+import net.engineeringdigest.journalApp.entity.JournalEntry;
+import net.engineeringdigest.journalApp.repository.UserRepository;
+import net.engineeringdigest.journalApp.entity.User;
 
 @Service
 public class JournalEntryService {
@@ -38,8 +38,7 @@ public class JournalEntryService {
                 throw new RuntimeException("Invalid user credentials");
             });
         } catch (RuntimeException runExcp) {
-            runExcp.printStackTrace(); // Log the error properly
-            throw new RuntimeException("Error creating journal entry", runExcp);
+            throw new RuntimeException(runExcp.getMessage(), runExcp.getCause());
         }
     }
 
@@ -51,19 +50,25 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
+    @Transactional
     public void remove(ObjectId id, String userName) {
         try {
             Optional<User> user = userService.findByUserName(userName);
             user.ifPresentOrElse(
                     userDetails -> {
-                        user.get().getJournalEntries().removeIf(journal -> journal.getId().equals(id));
-                        userService.createUser(user.get());
-                        journalEntryRepository.deleteById(id);
+                        boolean genuineJournal = userDetails.getJournalEntries().removeIf(journal -> journal.getId().equals(id));
+                        if(genuineJournal){
+                            userService.createUser(userDetails);
+                            journalEntryRepository.deleteById(id);
+                        }
+                        else{
+                            throw new RuntimeException("You are not right owner of the entry");
+                        }
                     }, () -> {
-                        throw new RuntimeException("Invalid user credentials");
+                        throw new UsernameNotFoundException("Invalid user credentials");
                     });
         } catch (RuntimeException e) {
-            throw new RuntimeException("Invalid user credentials");
+            throw new RuntimeException(e.getMessage(), e.getCause());
         }
     }
 }
